+++ l10n_it_einvoice_base/_check4deps_.py
@@ -9,14 +9,14 @@ import sys
 import re
 import logging
 
-from odoo import api, SUPERUSER_ID
-from odoo.exceptions import UserError
+from odoo import api, SUPERUSER_ID      # pylint: disable=import-error
+from odoo.exceptions import UserError   # pylint: disable=import-error
 
 _logger = logging.getLogger(__name__)
 
 
 def check_4_depending(cr):
-    """check_4_depending v2.0.20
+    """check_4_depending v2.0.24
     This function check for valid modules which current module depends on.
     Usually Odoo checks for depending on, through "depends" field in the manifest, but
     Odoo does not check for the version range neither check for incompatibilities.
@@ -255,9 +255,3 @@ def check_4_depending(cr):
     check_for_all_dependecies(
         manifest.get("conflicts", []), mtype="conflicts", disable_check=disable_check
     )
-
-
-
-
-
-