+++ l10n_it_einvoice_out/__manifest__.py
@@ -9,7 +9,7 @@
 #
 {
     "name": "Italian Localization - FatturaPA - Emissione",
-    "version": "10.0.1.0.30",
+    "version": "10.0.1.0.31",
     "category": "Localization/Italy",
     "summary": "E-Invoice emission",
     "author": "SHS-AV s.r.l.",