+++ l10n_it_einvoice_out_rc/models/__init__.py
@@ -1,3 +1,4 @@
 # -*- coding: utf-8 -*-
 from . import rc_type
+from . import account_fiscal_position
 from . import account_invoice