+++ l10n_it_einvoice_out_rc/models/account_fiscal_position.py
@@ -1,0 +1,12 @@
+# -*- coding: utf-8 -*-
+
+from odoo import fields, models
+
+
+class AccountFiscalPosition(models.Model):
+    _inherit = "account.fiscal.position"
+
+    rc_fiscal_document_type_id = fields.Many2one(
+        "italy.ade.invoice.type",
+        string="Self Invoice Fiscal Document Type",
+    )