+++ l10n_it_einvoice_out_rc/models/rc_type.py
@@ -5,6 +5,7 @@ from odoo import models, fields
 class AccountRCType(models.Model):
     _inherit = "account.rc.type"
 
+    # Deprecated
     fiscal_document_type_id = fields.Many2one(
         "italy.ade.invoice.type",
         string="Fiscal Document Type",