+++ l10n_it_einvoice_out_rc/wizard/wizard_export_e_invoice.py
@@ -63,8 +63,7 @@ class WizardExportFatturapa(models.TransientModel):
             invoice.rc_purchase_invoice_id and
             invoice.rc_purchase_invoice_id.fiscal_position_id and
             invoice.rc_purchase_invoice_id.fiscal_position_id.rc_type_id and
-            invoice.rc_purchase_invoice_id.fiscal_position_id.rc_type_id.
-                fiscal_document_type_id
+            invoice.fiscal_document_type_id
         ):
             body.DatiGenerali.DatiGeneraliDocumento.TipoDocumento = (
                 invoice.fiscal_document_type_id.code)